test: implement TDD backlog tests in correct projects and finalize XPlat coverage pipeline
