+++ tests/NexJob.Tests/FutureCoreHardeningTests.cs
@@ -1,0 +1,31 @@
+using FluentAssertions;
+using Xunit;
+
+namespace NexJob.Tests;
+
+/// <summary>
+/// TDD tests representing expected behaviors for items in TECH_DEBT_BACKLOG.md.
+/// These tests are currently skipped because they require business logic changes.
+/// </summary>
+public sealed class FutureCoreHardeningTests
+{
+    /// <summary>
+    /// TD003: Trigger should support complex input types directly.
+    /// Expected: Trigger can be configured with a specific TInput type.
+    /// </summary>
+    [Fact(Skip = "TD003: Broker Trigger Input Type Generalization. Requires architectural change in Trigger Options.")]
+    public void Trigger_ShouldSupportComplexInputTypes_Directly()
+    {
+        true.Should().BeTrue("Architectural validation placeholder");
+    }
+
+    /// <summary>
+    /// TD004: JobExecutor should have an interface for better testability.
+    /// Expected: We can mock IJobExecutor in JobDispatcherService tests.
+    /// </summary>
+    [Fact(Skip = "TD004: JobExecutor Testability. Requires extraction of IJobExecutor interface.")]
+    public void Dispatcher_ShouldUseJobExecutorInterface()
+    {
+        true.Should().BeTrue("Design validation placeholder");
+    }
+}
