+++ tests/NexJob.Trigger.AwsSqs.Tests/SqsFutureHardeningTests.cs
@@ -1,0 +1,23 @@
+using FluentAssertions;
+using Xunit;
+
+namespace NexJob.Trigger.AwsSqs.Tests;
+
+/// <summary>
+/// TDD tests representing expected behaviors for items in TECH_DEBT_BACKLOG.md.
+/// These tests are currently skipped because they require business logic changes.
+/// </summary>
+public sealed class SqsFutureHardeningTests
+{
+    /// <summary>
+    /// TD001: AWS SQS visibility extension should not overlap with retry visibility.
+    /// Expected: When enqueue fails, visibility extension loop stops immediately.
+    /// </summary>
+    [Fact(Skip = "TD001: AWS SQS Visibility Flakiness. Requires audit of visibility extension loop cancellation.")]
+    public async Task Sqs_EnqueueFailure_ShouldStopVisibilityExtensionImmediately()
+    {
+        // Verification logic for visibility cancellation
+        await Task.CompletedTask;
+        true.Should().BeTrue();
+    }
+}
