+++ tests/NexJob.Trigger.Kafka.Tests/KafkaFutureHardeningTests.cs
@@ -5,17 +5,15 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 using Moq;
 using NexJob.Internal;
-using NexJob.Trigger.Kafka;
 using Xunit;
 
-namespace NexJob.Tests;
+namespace NexJob.Trigger.Kafka.Tests;
 
 /// <summary>
-/// Tests representing expected behaviors for items in TECH_DEBT_BACKLOG.md.
+/// TDD tests representing expected behaviors for items in TECH_DEBT_BACKLOG.md.
 /// These tests are currently skipped because they require business logic changes.
-/// Activating them is the goal of the next refactoring phase.
 /// </summary>
-public sealed class FutureReliabilityTests
+public sealed class KafkaFutureHardeningTests
 {
     /// <summary>
     /// TD002: Kafka message without job_type header should be moved to DLT.
@@ -34,7 +32,6 @@ public sealed class FutureReliabilityTests
             ConsumeTimeout = TimeSpan.FromMilliseconds(10),
         };
 
-        // Create message without nexjob.job_type header
         var result = new ConsumeResult<string, string>
         {
             Message = new Message<string, string> { Key = "k1", Value = "{}", Headers = new Headers() },
@@ -55,45 +52,7 @@ public sealed class FutureReliabilityTests
         await (Task)method!.Invoke(sut, new object[] { result, CancellationToken.None })!;
 
         // Assert
-        // The message should be moved to DLT even if job_type is missing
         consumerMock.Verify(x => x.ProduceToDeadLetterAsync("test-dlt", result, It.Is<Exception>(e => e is InvalidOperationException), It.IsAny<CancellationToken>()), Times.Once);
         consumerMock.Verify(x => x.Commit(result), Times.Once);
-    }
-
-    /// <summary>
-    /// TD003: Trigger should support complex input types directly.
-    /// Expected: We can define a job with IJob&lt;MyModel&gt; and the trigger enqueues it correctly.
-    /// </summary>
-    [Fact(Skip = "TD003: Broker Trigger Input Type Generalization. Requires architectural change in Trigger Options.")]
-    public void Trigger_ShouldSupportComplexInputTypes_Directly()
-    {
-        // This is a design test. It would involve changing the TriggerOptions 
-        // to accept a generic TInput or a Type.
-        // For now, we mock the intention:
-        
-        /* 
-        var options = new RabbitMqTriggerOptions { 
-            InputType = typeof(MyModel) 
-        };
-        ...
-        scheduler.Verify(x => x.EnqueueAsync(It.Is<JobRecord>(j => j.InputType == typeof(MyModel).FullName)));
-        */
-        
-        true.Should().BeTrue("Placeholder for architectural validation");
-    }
-
-    /// <summary>
-    /// TD001: AWS SQS visibility extension should not overlap with retry visibility.
-    /// Expected: When enqueue fails, visibility extension loop stops immediately.
-    /// </summary>
-    [Fact(Skip = "TD001: AWS SQS Visibility Flakiness. Requires audit of visibility extension loop cancellation.")]
-    public async Task Sqs_EnqueueFailure_ShouldStopVisibilityExtensionImmediately()
-    {
-        // This test would verify that the extension task is cancelled 
-        // BEFORE the method exits on failure, preventing it from extending 
-        // visibility of a message that was just "returned" to the queue for retry.
-        
-        await Task.CompletedTask;
-        true.Should().BeTrue();
     }
 }
